requirements.txt
django==5.2.*
psycopg2-binary==2.9.*
python-dotenv
